pyproject.toml
[tool.poetry]
name = "ai-powered chat application"
version = "0.1.0"
description = "Simple interactive AI chat"
authors = ["Shahriar Shariati <shahriarshm81@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic = "^2.4.2"
tortoise-orm = {version = "^0.20.0", extras = ["asyncpg"]}
uvicorn = "^0.23.2"
black = "^23.10.1"
isort = "^5.12.0"
g4f = "^0.1.7.9"
aerich = "^0.7.2"
python-dotenv = "^1.0.0"

[tool.aerich]
tortoise_orm = "src.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
